+++ angular4-seed/config/webpack.common.js
@@ -1,10 +1,11 @@
 const webpack = require('webpack');
 const helpers = require('./helpers');
 
-/*
+/**
  * Webpack Plugins
+ *
+ * problem with copy-webpack-plugin
  */
-// problem with copy-webpack-plugin
 const AssetsPlugin = require('assets-webpack-plugin');
 const NormalModuleReplacementPlugin = require('webpack/lib/NormalModuleReplacementPlugin');
 const ContextReplacementPlugin = require('webpack/lib/ContextReplacementPlugin');
@@ -13,26 +14,30 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 const CheckerPlugin = require('awesome-typescript-loader').CheckerPlugin;
 const HtmlElementsPlugin = require('./html-elements-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const InlineManifestWebpackPlugin = require('inline-manifest-webpack-plugin');
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
 const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin');
+const ngcWebpack = require('ngc-webpack');
+//const PreloadWebpackPlugin = require('preload-webpack-plugin');
 
-/*
+/**
  * Webpack Constants
  */
 const HMR = helpers.hasProcessFlag('hot');
-const AOT = helpers.hasNpmFlag('aot');
+const AOT = process.env.BUILD_AOT || helpers.hasNpmFlag('aot');
 const METADATA = {
-  title: 'Light Blue 3.7.0 with Angular 4.0 Final Release support by Flatlogic',
+  title: 'Light Blue 3.8.0 with Angular 4.0 Final Release support by Flatlogic',
   baseUrl: '/',
-  isDevServer: helpers.isWebpackDevServer()
+  isDevServer: helpers.isWebpackDevServer(),
+  HMR: HMR
 };
 
-/*
+/**
  * Webpack configuration
  *
  * See: http://webpack.github.io/docs/configuration.html#cli
  */
-module.exports = function(options) {
+module.exports = function (options) {
   isProd = options.env === 'production';
   return {
     performance: { hints: false },
@@ -43,75 +48,102 @@ module.exports = function(options) {
      *
      * See: http://webpack.github.io/docs/configuration.html#cache
      */
-     //cache: false,
+    //cache: false,
 
-    /*
+    /**
      * The entry point for the bundle
      * Our Angular.js app
      *
      * See: http://webpack.github.io/docs/configuration.html#entry
      */
     entry: {
+
       'polyfills': './src/polyfills.browser.ts',
       'vendor':    './src/vendor.browser.ts',
       'main':      AOT ? './src/main.browser.aot.ts' :
-        './src/main.browser.ts'
+                  './src/main.browser.ts'
+
     },
 
-    /*
+    /**
      * Options affecting the resolving of modules.
      *
      * See: http://webpack.github.io/docs/configuration.html#resolve
      */
     resolve: {
 
-      /*
+      /**
        * An array of extensions that should be used to resolve modules.
        *
        * See: http://webpack.github.io/docs/configuration.html#resolve-extensions
        */
       extensions: ['.ts', '.js', '.json'],
 
-      // An array of directory names to be resolved to the current directory
-      modules: [helpers.root('src'), helpers.root('node_modules')]
+      /**
+       * An array of directory names to be resolved to the current directory
+       */
+      modules: [helpers.root('src'), helpers.root('node_modules')],
 
     },
 
-    /*
+    /**
      * Options affecting the normal modules.
      *
      * See: http://webpack.github.io/docs/configuration.html#module
      */
     module: {
 
-      /*
-       * An array of automatically applied loaders.
-       *
-       * IMPORTANT: The loaders here are resolved relative to the resource which they are applied to.
-       * This means they are not resolved relative to the configuration file.
-       *
-       * See: http://webpack.github.io/docs/configuration.html#module-loaders
-       */
       rules: [
-        /*
-         * Typescript loader support for .ts and Angular 2 async routes via .async.ts
-         * Replace templateUrl and stylesUrl with require()
+
+        /**
+         * Typescript loader support for .ts
+         *
+         * Component Template/Style integration using `angular2-template-loader`
+         * Angular 2 lazy loading (async routes) via `ng-router-loader`
+         *
+         * `ng-router-loader` expects vanilla JavaScript code, not TypeScript code. This is why the
+         * order of the loader matter.
          *
          * See: https://github.com/s-panferov/awesome-typescript-loader
          * See: https://github.com/TheLarkInn/angular2-template-loader
+         * See: https://github.com/shlomiassaf/ng-router-loader
          */
         {
           test: /\.ts$/,
           use: [
-            '@angularclass/hmr-loader?pretty=' + !isProd + '&prod=' + isProd,
-            'awesome-typescript-loader?{tsconfig: "tsconfig.webpack.json"}',
-            'angular2-template-loader',
-            'angular-router-loader?loader=system&genDir=compiled/src/app&aot=' + AOT
+            {
+              loader: '@angularclass/hmr-loader',
+              options: {
+                pretty: !isProd,
+                prod: isProd
+              }
+            },
+            {
+              /**
+               *  MAKE SURE TO CHAIN VANILLA JS CODE, I.E. TS COMPILATION OUTPUT.
+               */
+              loader: 'ng-router-loader',
+              options: {
+                loader: 'async-import',
+                genDir: 'compiled',
+                aot: AOT
+              }
+            },
+            {
+              loader: 'awesome-typescript-loader',
+              options: {
+                tsconfig: 'tsconfig.webpack.json',
+                useCache: !isProd
+              }
+            },
+            {
+              loader: 'angular2-template-loader'
+            }
           ],
           exclude: [/\.(spec|e2e)\.ts$/]
         },
 
-        /*
+        /**
          * Json loader support for *.json files.
          *
          * See: https://github.com/webpack/json-loader
@@ -149,35 +181,39 @@ module.exports = function(options) {
           exclude: [helpers.root('src/index.html')]
         },
 
-        /* File loader for supporting images, for example, in CSS files.
-        */
+        /**
+         * File loader for supporting images, for example, in CSS files.
+         */
         {
           test: /\.(jpg|png|gif)$/,
           use: 'file-loader'
         }
-      ]
+
+      ],
+
     },
 
-    /*
+    /**
      * Add additional plugins to the compiler.
      *
      * See: http://webpack.github.io/docs/configuration.html#plugins
      */
     plugins: [
-      new AssetsPlugin({
-        path: helpers.root('dist'),
-        filename: 'webpack-assets.json',
-        prettyPrint: true
-      }),
+      // Use for DLLs
+      // new AssetsPlugin({
+      //   path: helpers.root('dist'),
+      //   filename: 'webpack-assets.json',
+      //   prettyPrint: true
+      // }),
 
-      /*
+      /**
        * Plugin: ForkCheckerPlugin
        * Description: Do type checking in a separate process, so webpack don't need to wait.
        *
        * See: https://github.com/s-panferov/awesome-typescript-loader#forkchecker-boolean-defaultfalse
        */
       new CheckerPlugin(),
-      /*
+      /**
        * Plugin: CommonsChunkPlugin
        * Description: Shares common code between the pages.
        * It identifies common modules and put them into a commons chunk.
@@ -188,6 +224,24 @@ module.exports = function(options) {
       new CommonsChunkPlugin({
         name: ['polyfills', 'vendor'].reverse()
       }),
+      /**
+       * This enables tree shaking of the vendor modules
+       */
+      // new CommonsChunkPlugin({
+      //   name: 'vendor',
+      //   chunks: ['main'],
+      //   minChunks: module => /node_modules/.test(module.resource)
+      // }),
+      /**
+       * Specify the correct order the scripts will be injected in
+       */
+      // new CommonsChunkPlugin({
+      //   name: ['polyfills', 'vendor'].reverse()
+      // }),
+      // new CommonsChunkPlugin({
+      //   name: ['manifest'],
+      //   minChunks: Infinity,
+      // }),
 
       /**
        * Plugin: ContextReplacementPlugin
@@ -197,15 +251,19 @@ module.exports = function(options) {
        * See: https://github.com/angular/angular/issues/11580
        */
       new ContextReplacementPlugin(
-        // The (\\|\/) piece accounts for path separators in *nix and Windows
-        /angular(\\|\/)core(\\|\/)src(\\|\/)linker/,
+        /**
+         * The (\\|\/) piece accounts for path separators in *nix and Windows
+         */
+        /angular(\\|\/)core(\\|\/)@angular/,
         helpers.root('src'), // location of your src
         {
-          // your Angular Async Route paths relative to this root directory
+          /**
+           * Your Angular Async Route paths relative to this root directory
+           */
         }
       ),
 
-      /*
+      /**
        * Plugin: CopyWebpackPlugin
        * Description: Copy files and directories in webpack.
        *
@@ -219,13 +277,13 @@ module.exports = function(options) {
       }]),
 
       /*
-       * Plugin: HtmlWebpackPlugin
-       * Description: Simplifies creation of HTML files to serve your webpack bundles.
-       * This is especially useful for webpack bundles that include a hash in the filename
-       * which changes every compilation.
-       *
-       * See: https://github.com/ampedandwired/html-webpack-plugin
-       */
+      * Plugin: HtmlWebpackPlugin
+      * Description: Simplifies creation of HTML files to serve your webpack bundles.
+      * This is especially useful for webpack bundles that include a hash in the filename
+      * which changes every compilation.
+      *
+      * See: https://github.com/ampedandwired/html-webpack-plugin
+      */
       new HtmlWebpackPlugin({
         template: 'src/index.html',
         title: METADATA.title,
@@ -285,32 +343,58 @@ module.exports = function(options) {
         headTags: require('./head-config.common')
       }),
 
+      /**
+       * Plugin LoaderOptionsPlugin (experimental)
+       *
+       * See: https://gist.github.com/sokra/27b24881210b56bbaff7
+       */
       new LoaderOptionsPlugin({}),
 
-      // Fix Angular 2
-      new NormalModuleReplacementPlugin(
-        /facade(\\|\/)async/,
-        helpers.root('node_modules/@angular/core/src/facade/async.js')
-      ),
-      new NormalModuleReplacementPlugin(
-        /facade(\\|\/)collection/,
-        helpers.root('node_modules/@angular/core/src/facade/collection.js')
-      ),
-      new NormalModuleReplacementPlugin(
-        /facade(\\|\/)errors/,
-        helpers.root('node_modules/@angular/core/src/facade/errors.js')
-      ),
-      new NormalModuleReplacementPlugin(
-        /facade(\\|\/)lang/,
-        helpers.root('node_modules/@angular/core/src/facade/lang.js')
-      ),
-      new NormalModuleReplacementPlugin(
-        /facade(\\|\/)math/,
-        helpers.root('node_modules/@angular/core/src/facade/math.js')
-      )
+      new ngcWebpack.NgcWebpackPlugin({
+        /**
+         * If false the plugin is a ghost, it will not perform any action.
+         * This property can be used to trigger AOT on/off depending on your build target (prod, staging etc...)
+         *
+         * The state can not change after initializing the plugin.
+         * @default true
+         */
+        disabled: !AOT,
+        tsConfig: helpers.root('tsconfig.webpack.json'),
+        /**
+         * A path to a file (resource) that will replace all resource referenced in @Components.
+         * For each `@Component` the AOT compiler compiles it creates new representation for the templates (html, styles)
+         * of that `@Components`. It means that there is no need for the source templates, they take a lot of
+         * space and they will be replaced by the content of this resource.
+         *
+         * To leave the template as is set to a falsy value (the default).
+         *
+         * TIP: Use an empty file as an overriding resource. It is recommended to use a ".js" file which
+         * usually has small amount of loaders hence less performance impact.
+         *
+         * > This feature is doing NormalModuleReplacementPlugin for AOT compiled resources.
+         *
+         * ### resourceOverride and assets
+         * If you reference assets in your styles/html that are not inlined and you expect a loader (e.g. url-loader)
+         * to copy them, don't use the `resourceOverride` feature as it does not support this feature at the moment.
+         * With `resourceOverride` the end result is that webpack will replace the asset with an href to the public
+         * assets folder but it will not copy the files. This happens because the replacement is done in the AOT compilation
+         * phase but in the bundling it won't happen (it's being replaced with and empty file...)
+         *
+         * @default undefined
+         */
+        resourceOverride: helpers.root('config/resource-override.js')
+      }),
+
+      /**
+       * Plugin: InlineManifestWebpackPlugin
+       * Inline Webpack's manifest.js in index.html
+       *
+       * https://github.com/szrenwei/inline-manifest-webpack-plugin
+       */
+      new InlineManifestWebpackPlugin(),
     ],
 
-    /*
+    /**
      * Include polyfills or mocks for various node stuff
      * Description: Node configuration
      *