+++ angular4-seed/src/main.browser.aot.ts
@@ -1,23 +1,34 @@
+/**
+ * Angular bootstrapping
+ */
 import { platformBrowser } from '@angular/platform-browser';
 import { decorateModuleRef } from './app/environment';
-/*
+/**
  * App Module
- * our top level module that holds all of our components
+ * our top level module that holds all of our components.
  */
 import { AppModuleNgFactory } from '../compiled/src/app/app.module.ngfactory';
-
-/*
- * Bootstrap our Angular app with a top level NgModule
+/**
+ * Bootstrap our Angular app with a top level NgModule.
  */
 export function main(): Promise<any> {
   return platformBrowser()
     .bootstrapModuleFactory(AppModuleNgFactory)
     .then(decorateModuleRef)
-    .catch(err => console.error(err));
+    .catch((err) => console.error(err));
 }
 
-export function bootstrapDomReady() {
-  document.addEventListener('DOMContentLoaded', main);
+switch (document.readyState) {
+  case 'loading':
+    document.addEventListener('DOMContentLoaded', _domReadyHandler, false);
+    break;
+  case 'interactive':
+  case 'complete':
+  default:
+    main();
 }
 
-bootstrapDomReady();
+function _domReadyHandler() {
+  document.removeEventListener('DOMContentLoaded', _domReadyHandler, false);
+  main();
+}