+++ angular4-seed/src/vendor.browser.ts
@@ -1,18 +1,3 @@
-import '@angular/platform-browser';
-import '@angular/platform-browser-dynamic';
-import '@angular/core';
-import '@angular/common';
-import '@angular/forms';
-import '@angular/http';
-import '@angular/router';
-
-// AngularClass
-import '@angularclass/hmr';
-
-// RxJS
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mergeMap';
-
 import 'jquery';
 import 'tether';
 import 'bootstrap';