+++ angular4/src/app/special/special.module.ts
@@ -6,8 +6,6 @@ import { RouterModule } from '@angular/router';
 import { Search } from './search/search.component';
 import { Invoice } from './invoice/invoice.component';
 
-console.log('`Detail` bundle loaded asynchronously');
-// async components must be named routes for WebpackAsyncRoute
 export const routes = [
   {path: '', redirectTo: 'search', pathMatch: 'full'},
   {path: 'search', component: Search},